+++ dcp-common/src/test/java/com/vmware/dcp/common/test/VerificationHost.java
@@ -96,6 +96,7 @@ import com.vmware.dcp.services.common.QueryValidationTestService.QueryValidation
 import com.vmware.dcp.services.common.ServiceHostLogService.LogServiceState;
 import com.vmware.dcp.services.common.ServiceHostManagementService;
 import com.vmware.dcp.services.common.ServiceUriPaths;
+import com.vmware.dcp.services.common.UserService.UserState;
 
 public class VerificationHost extends ExampleServiceHost {
     private static final int LARGE_PAYLOAD_SIZE_BYTES = 1024 * 512;
@@ -2526,6 +2527,30 @@ public class VerificationHost extends ExampleServiceHost {
 
         // Associate resulting authorization context with this thread
         setAuthorizationContext(ab.getResult());
+    }
+
+    public String createUserService(String email) throws Throwable {
+        final String[] userUriPath = new String[1];
+
+        UserState userState = new UserState();
+        userState.email = email;
+
+        URI postUserUri = UriUtils.buildUri(this, ServiceUriPaths.CORE_AUTHZ_USERS);
+        testStart(1);
+        send(Operation
+                .createPost(postUserUri)
+                .setBody(userState)
+                .setCompletion((o, e) -> {
+                    if (e != null) {
+                        failIteration(e);
+                        return;
+                    }
+                    UserState state = o.getBody(UserState.class);
+                    userUriPath[0] = state.documentSelfLink;
+                    completeIteration();
+                }));
+        testWait();
+        return userUriPath[0];
     }
 
     public void deleteAllChildServices(URI factoryURI) throws Throwable {